backend/data-service/requirements.txt
flask
flask_cors
cassandra-driver==3.25.0
python-dotenv
requests